Módulo 6 - Explorando mais do JPA e Hibernate

+++ algafood-api/src/main/java/com/algaworks/algafood/domain/service/CadastroEstadoService.java
@@ -2,7 +2,6 @@ package com.algaworks.algafood.domain.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.algaworks.algafood.domain.exception.EntidadeEmUsoException;
@@ -12,26 +11,25 @@ import com.algaworks.algafood.domain.repository.EstadoRepository;
 
 @Service
 public class CadastroEstadoService {
-
-	@Autowired
-    private EstadoRepository estadoRepository;
 	
-	 public Estado salvar(Estado estado) {
-         return estadoRepository.save(estado);
-     }
-	 
-	 public void excluir(Long estadoId) {
-         try {
-             estadoRepository.deleteById(estadoId);
-             
-         } catch (EmptyResultDataAccessException e) {
-             throw new EntidadeNaoEncontradaException(
-                 String.format("Não existe um cadastro de estado com código %d", estadoId));
-         
-         } catch (DataIntegrityViolationException e) {
-             throw new EntidadeEmUsoException(
-                 String.format("Estado de código %d não pode ser removido, pois está em uso", estadoId));
-         }
-     }
-     
+	@Autowired
+	private EstadoRepository estadoRepository;
+
+	public Estado salvar(Estado estado) {
+		return estadoRepository.save(estado);
+	}
+
+	public void excluir(Long estadoId) {
+		try {
+			if(!estadoRepository.existsById(estadoId)) {
+	    		throw new EntidadeNaoEncontradaException(
+	    				String.format("Não existe um cadastro de estado com código %d", estadoId));
+	        }
+			estadoRepository.deleteById(estadoId);
+			
+		} catch (DataIntegrityViolationException e) {
+			throw new EntidadeEmUsoException(
+					String.format("Estado de código %d não pode ser removido, pois está em uso", estadoId));
+		}
+	}
 }
